+++ src/main/java/identification/Identification.java
@@ -1,4 +1,4 @@
-package core;
+package identification;
 
 import listener.RawOdds;
 import listener.RawOddsListener;
